+++ src/components/icon/icon-unicodes.js
@@ -299,9 +299,7 @@ const legacyNames = {
   // it tries to use `delete` as an identifier, but `delete` is a reserved keyword.
   // By using ["delete"] instead, tsc generates a different - valid - .d.ts file
   // See also: https://github.com/microsoft/TypeScript/issues/53111
-  /* eslint-disable-next-line dot-notation */
   bin: batchActions["delete"],
-  /* eslint-disable-next-line dot-notation */
   bulk_destroy: batchActions["delete"],
   caret_down: actions.dropdown,
   collaborate: actions.share,