+++ src/components/search/search.component.tsx
@@ -9,7 +9,6 @@ import Icon from "../icon";
 import Textbox from "../textbox";
 import Button from "../button";
 import { ValidationProps } from "../../__internal__/validations";
-import Logger from "../../__internal__/utils/logger";
 import useLocale from "../../hooks/__internal__/useLocale";
 import Events from "../../__internal__/utils/helpers/events";
 
@@ -26,8 +25,6 @@ export interface SearchProps extends ValidationProps, MarginProps, TagProps {
   "aria-label"?: string;
   /** Prop to specify the aria-label of the search button */
   searchButtonAriaLabel?: string;
-  /** Prop for `uncontrolled` use */
-  defaultValue?: string;
   /** Prop for `id` */
   id?: string;
   /** Prop for `name` */
@@ -35,7 +32,7 @@ export interface SearchProps extends ValidationProps, MarginProps, TagProps {
   /** Prop for `onBlur` events */
   onBlur?: (ev: React.FocusEvent<HTMLInputElement>) => void;
   /** Prop for `onChange` events */
-  onChange?: (ev: SearchEvent) => void;
+  onChange: (ev: SearchEvent) => void;
   /** Prop for `onClick` events.
    *  `onClick` events are triggered when the `searchButton` is clicked
    */
@@ -66,8 +63,8 @@ export interface SearchProps extends ValidationProps, MarginProps, TagProps {
    * Leaving the `maxWidth` prop with no value will default the width to '100%'
    */
   maxWidth?: string;
-  /** Prop for `controlled` use */
-  value?: string;
+  /** Current value */
+  value: string;
   /** Prop to specify the styling of the search component */
   variant?: "default" | "dark";
   /** Input tabindex */
@@ -81,12 +78,9 @@ export type SearchHandle = {
   focus: () => void;
 } | null;
 
-let deprecateUncontrolledWarnTriggered = false;
-
 export const Search = React.forwardRef<SearchHandle, SearchProps>(
   (
     {
-      defaultValue,
       onChange,
       onClick,
       onFocus,
@@ -112,8 +106,6 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
     },
     ref,
   ) => {
-    const isControlled = value !== undefined;
-    const initialValue = isControlled ? value : defaultValue;
     const locale = useLocale();
     const searchRef = useRef<HTMLDivElement>(null);
     const inputRef = useRef<HTMLInputElement>(null);
@@ -128,35 +120,16 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
       [],
     );
 
-    if (!deprecateUncontrolledWarnTriggered && !isControlled) {
-      deprecateUncontrolledWarnTriggered = true;
-      Logger.deprecate(
-        "Uncontrolled behaviour in `Search` is deprecated and support will soon be removed. Please make sure all your inputs are controlled.",
-      );
-    }
+    invariant(typeof value === "string", "This component has no initial value");
 
-    invariant(
-      typeof initialValue === "string",
-      "This component has no initial value",
-    );
-
-    const [searchValue, setSearchValue] = useState(initialValue);
     const [isFocused, setIsFocused] = useState(false);
 
-    const isSearchValueEmpty = !isControlled
-      ? searchValue.length === 0
-      : value.length === 0;
+    const isSearchValueEmpty = value.length === 0;
 
     let buttonProps = {};
 
     const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      if (onChange) {
-        onChange(event);
-      }
-
-      if (!isControlled) {
-        setSearchValue(event.target.value);
-      }
+      onChange(event);
     };
 
     const handleFocus = (event: React.FocusEvent<HTMLInputElement>) => {
@@ -173,7 +146,7 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
             target: {
               name,
               id,
-              value: !isControlled ? searchValue : value,
+              value,
             },
           });
         },
@@ -181,8 +154,6 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
     }
 
     const handleIconClick = () => {
-      setSearchValue("");
-
       onChange?.({
         target: {
           ...(name && { name }),
@@ -224,7 +195,7 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
 
       if (Events.isEscKey(event) && !isSearchValueEmpty) {
         event.stopPropagation();
-        setSearchValue("");
+
         onChange?.({
           target: {
             ...(name && { name }),
@@ -243,9 +214,7 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
       typeof searchButton === "string"
         ? searchButton
         : locale.search.searchButtonText();
-    const searchHasValue = !isControlled
-      ? !!searchValue?.length
-      : !!value?.length;
+    const searchHasValue = !!value?.length;
 
     return (
       <StyledSearch
@@ -264,7 +233,7 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
       >
         <Textbox
           placeholder={placeholder}
-          value={!isControlled ? searchValue : value}
+          value={value}
           inputIcon={!isSearchValueEmpty ? "cross" : undefined}
           iconTabIndex={!isSearchValueEmpty ? 0 : -1}
           iconOnClick={handleIconClick}
@@ -305,7 +274,5 @@ export const Search = React.forwardRef<SearchHandle, SearchProps>(
     );
   },
 );
-
-Search.displayName = "Search";
 
 export default Search;